+++ sqlite_crdt/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.8
+
+- Dedicated node_id column to improve query performance
+- Add method to reset node id
+- Improve merge error reporting
+
 ## 0.0.7
 
 * Move database-agnostic code to sql_crdt